+++ docs/changelog.rst
@@ -27,6 +27,7 @@ news
   special graphs stretched the x-axis to equalize the x-axis from other
   variable for do the mosaic of inspections of series in EDA
 - New: Make the matrix plots of scatter plots of all series of the var D
+- New: Climatology table and graphs for var D
 
 fixes
 +++++